refactor: DH-23199: Move deep Python-to-Java conversion into jcompat.j_json_value
The recursive container conversion used for js plugin loader details is generic, so promote it from
the private deephaven_internal.plugin.js._to_j_object to a public deephaven.jcompat.j_json_value.
It behaves like j_hashmap but converts nested containers, and now also unwraps JObjectWrapper
values. Tests move from test_plugin_js to test_jcompat.
