+++ py/server/deephaven/jcompat.py
@@ -49,6 +49,8 @@ _JPrimitiveArrayConversionUtility = jpy.get_type(
     "io.deephaven.integrations.common.PrimitiveArrayConversionUtility"
 )
 _JTableDefinition = jpy.get_type("io.deephaven.engine.table.TableDefinition")
+_JHashMap = jpy.get_type("java.util.HashMap")
+_JArrayList = jpy.get_type("java.util.ArrayList")
 
 _DH_PANDAS_NULLABLE_TYPE_MAP: dict[DType, type[ExtensionDtype]] = {
     bool_: pd.BooleanDtype,
@@ -88,6 +90,29 @@ def j_hashmap(d: Optional[dict] = None) -> Optional[jpy.JType]:
         v = unwrap(v)
         r.put(k, v)
     return r
+
+
+def j_json_value(value: Any) -> Any:
+    """Recursively converts a JSON-compatible Python value into Java objects.
+
+    A dict is converted to a java.util.HashMap and a list, tuple, set, or frozenset is converted to a
+    java.util.ArrayList, with their keys and values converted recursively. Any other value is unwrapped and
+    left for jpy to convert.
+
+    Nested containers must be converted explicitly; otherwise jpy passes them through as opaque
+    org.jpy.PyObject values.
+    """
+    if isinstance(value, dict):
+        r = _JHashMap(len(value))
+        for k, v in value.items():
+            r.put(j_json_value(k), j_json_value(v))
+        return r
+    if isinstance(value, (list, tuple, set, frozenset)):
+        r = _JArrayList(len(value))
+        for v in value:
+            r.add(j_json_value(v))
+        return r
+    return unwrap(value)
 
 
 def j_hashset(s: Optional[Union[set, Sequence]] = None) -> Optional[jpy.JType]:
